Osdag crashes when reset button was clicked and is fixed

+++ Connections/Moment/BBSpliceCoverPlate/BBSpliceCoverPlateBolted/coverplate_bolted_main.py
@@ -148,7 +148,7 @@ class DesignPreferences(QDialog):
 		designPref["bolt"] = {}
 		designPref["bolt"]["bolt_hole_type"] = str(self.ui.combo_boltHoleType.currentText())
 		designPref["bolt"]["bolt_hole_clrnce"] = self.get_clearance()
-		designPref["bolt"]["bolt_fu"] = int(self.ui.txt_boltFu.text())
+		designPref["bolt"]["bolt_fu"] = float(self.ui.txt_boltFu.text())
 		self.ui.combo_slipfactor.setCurrentIndex(4)
 		designPref["bolt"]["slip_factor"] = float(str(self.ui.combo_slipfactor.currentText()))
